+++ src/backend/columnar/cstore_reader.c
@@ -34,8 +34,8 @@
 #include "utils/lsyscache.h"
 #include "utils/rel.h"
 
-#include "cstore.h"
-#include "cstore_version_compat.h"
+#include "columnar/cstore.h"
+#include "columnar/cstore_version_compat.h"
 
 /* static function declarations */
 static StripeBuffers * LoadFilteredStripeBuffers(Relation relation,
