+++ src/backend/columnar/cstore_writer.c
@@ -24,8 +24,8 @@
 #include "utils/memutils.h"
 #include "utils/rel.h"
 
-#include "cstore.h"
-#include "cstore_version_compat.h"
+#include "columnar/cstore.h"
+#include "columnar/cstore_version_compat.h"
 
 static StripeBuffers * CreateEmptyStripeBuffers(uint32 stripeMaxRowCount,
 												uint32 blockRowCount,
