+++ src/include/columnar/cstore_version_compat.h
